keos_project3::lazy_pager

Struct FileBackedLoader

pub struct FileBackedLoader {}

A loader for file-backed memory regions.

FileBackedLoader is used for memory mappings backed by files, such as when mmap is called with a regular file. This loader reads data from the underlying file starting at a specific offset and returns it in a newly allocated Page.

The offset within the file is determined based on the virtual address passed to load, relative to the mapping’s start address and offset.

This loader handles partial page reads and fills any unread bytes with zeroes.